rename(discount, menu): 패키지 test에서 main으로 이동
