+++ src/main/java/christmas/discount/DiscountPrice.java
@@ -1,4 +1,4 @@
-package christmas.domain.discount;
+package christmas.discount;
 
 public class DiscountPrice {
     private final int discountPriceValue;
